+++ packages/components-js/tests/vrt/specs/stepper-horizontal.vrt.ts
@@ -24,22 +24,28 @@ it('should have no visual regression for :hover + :focus-visible', async () => {
     await vrt.test('stepper-horizontal-states', async () => {
       const page = vrt.getPage();
 
-      const getElementsMarkup: GetThemedMarkup = (theme: Theme) => `
-        <p-stepper-horizontal theme="${theme}">
-          <p-stepper-horizontal-item theme="${theme}" state="warning">Step Warning</p-stepper-horizontal-item>
-          <p-stepper-horizontal-item theme="${theme}" state="complete">Step Complete</p-stepper-horizontal-item>
-          <p-stepper-horizontal-item theme="${theme}" state="warning" disabled>Step Warning Disabled</p-stepper-horizontal-item>
-          <p-stepper-horizontal-item theme="${theme}" state="complete" disabled>Step Complete Disabled</p-stepper-horizontal-item>
-          <p-stepper-horizontal-item theme="${theme}" state="current">Step Current</p-stepper-horizontal-item>
-          <p-stepper-horizontal-item theme="${theme}">Step</p-stepper-horizontal-item>
+      const stepperHorizontalItems = `
+        <p-stepper-horizontal-item state="warning">Warning</p-stepper-horizontal-item>
+        <p-stepper-horizontal-item state="complete">Complete</p-stepper-horizontal-item>
+        <p-stepper-horizontal-item state="warning" disabled>Warning Disabled</p-stepper-horizontal-item>
+        <p-stepper-horizontal-item state="complete" disabled>Complete Disabled</p-stepper-horizontal-item>
+        <p-stepper-horizontal-item state="current">Current</p-stepper-horizontal-item>
+        <p-stepper-horizontal-item>Default</p-stepper-horizontal-item>`;
+
+      const getElementsMarkup: GetThemedMarkup = (theme: Theme) =>
+        `<p-stepper-horizontal theme="${theme}">
+          ${stepperHorizontalItems}
         </p-stepper-horizontal>
-        `;
 
-      await setContentWithDesignSystem(page, getThemedBodyMarkup(getElementsMarkup, { themes: ['light', 'dark'] }));
+        <p-stepper-horizontal theme="${theme}" size="medium">
+          ${stepperHorizontalItems}
+        </p-stepper-horizontal>`;
 
-      await forceHoverState(page, '.hover > p-stepper-horizontal p-stepper-horizontal-item >>> button');
-      await forceFocusState(page, '.focus > p-stepper-horizontal p-stepper-horizontal-item >>> button');
-      await forceFocusHoverState(page, '.focus-hover > p-stepper-horizontal p-stepper-horizontal-item  >>> button');
+      await setContentWithDesignSystem(page, getThemedBodyMarkup(getElementsMarkup));
+
+      await forceHoverState(page, '.hover p-stepper-horizontal p-stepper-horizontal-item >>> button');
+      await forceFocusState(page, '.focus p-stepper-horizontal p-stepper-horizontal-item >>> button');
+      await forceFocusHoverState(page, '.focus-hover p-stepper-horizontal p-stepper-horizontal-item >>> button');
     })
   ).toBeFalsy();
 });